Update macros.html.twig
Added {% import 'SyliusResourceBundle:Macros:buttons.html.twig' as buttons %} to macro tree. 

if i use this bundle. i will get the error that the variable buttons does not exists on line 38. After i added the import to the macro tree this problem was fixed. So I think that a import has to be at Macro level.